telegram-bot/requirements.txt
python-telegram-bot[job-queue]==21.0.1
confluent-kafka==2.4.0